+++ packages/fe-mockserver-core/src/request/applyParser.ts
@@ -333,7 +333,14 @@ export class ApplyParser extends FilterParser {
                 this.MANY_SEP({
                     SEP: COMMA,
                     DEF: () => {
-                        groupBy.push(this.CONSUME2(SIMPLEIDENTIFIER).image);
+                        const groupByStr: string[] = [];
+                        this.MANY_SEP2({
+                            SEP: SLASH,
+                            DEF: () => {
+                                groupByStr.push(this.CONSUME2(SIMPLEIDENTIFIER).image);
+                            }
+                        });
+                        groupBy.push(groupByStr.join('/'));
                     }
                 });
             });
@@ -377,6 +384,7 @@ export class ApplyParser extends FilterParser {
                         this.CONSUME5(WS);
                         this.CONSUME(AS_TOKEN);
                         this.CONSUME6(WS);
+                        // NetAmount%20with%20max%20as%20maxAmount
                         alias = this.CONSUME3(SIMPLEIDENTIFIER).image;
                     });